Eternal Quotex Bot Source/requirements.txt
PySide6>=6.7,<7
quotexpy>=1.40.7
beautifulsoup4==4.11.2
requests>=2.32
undetected-chromedriver>=3.5
websocket-client>=1.8
cryptography>=42.0
pandas>=2.2
scipy>=1.10
numpy>=1.24
Pillow>=10.4
playwright>=1.54
pyinstaller>=5.13,<6
python-dotenv>=1.0
